### 2023-11-13 #6056 @editorialbot
[PRE REVIEW]: G'MIC: An Open-Source Self-Extending Framework for Image Processing
**Submitting author:** @dtschump (<a href="http://orcid.org/0000-0003-3454-5079">David Tschumperlé</a>)
**Repository:** https://github.com/GreycLab/gmic
**Branch with paper.md** (empty if default branch): joss
**Version:** 3.3.2
