### 2023-11-13 #6056 @danielskatz
@dtschump - Can you add the full software to the joss branch (so that this is copy of master, with the paper, rather than just the paper)?
