### 2023-11-13 #6056 @danielskatz
Finally (for now), your paper is not a JOSS paper: it's about 7000 words long, while JOSS papers are typically around 1000 words. Much of what is in your paper might be replaced by links to documentation or something else. The JOSS paper should just be a brief introduction to the software.

Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)  You can use the command `@editorialbot generate pdf` after making changes to the .md file to make a new PDF.
