### 2023-11-13 0bb8816
Add entry for compiling JOSS paper from root Makefile.
