+++ Makefile
@@ -75,5 +75,10 @@
 
 all: all
 
+joss: paper.pdf
+
+paper.pdf: paper.md paper.bib
+	@docker run --rm --volume $$PWD:/data --user $$(id -u):$$(id -g) --env JOURNAL=joss openjournals/inara
+
 %:
 	cd src && $(MAKE) $*
