### 2023-11-13 1c64097
Add suggested DOIs in bibliography.
