+++ paper.bib
@@ -36,6 +36,7 @@
   booktitle={2022 IEEE International Conference on Image Processing (ICIP)},
   pages={371--375},
   year={2022},
+  doi = {10.1109/icip46576.2022.9897386},
   organization={IEEE}
 }
 
@@ -45,6 +46,7 @@
   booktitle={2022 IEEE International Conference on Image Processing (ICIP)},
   pages={3490--3494},
   year={2022},
+  doi={10.1109/icip46576.2022.9897334},
   organization={IEEE}
 }
 
@@ -63,6 +65,7 @@
   number={3},
   pages={1511--1535},
   year={2020},
+  doi={10.1137/19m1306798},
   publisher={SIAM}
 }
 
@@ -74,6 +77,7 @@
   number={6},
   pages={1809--1824},
   year={2015},
+  doi={10.1109/tip.2015.2411437},
   publisher={IEEE}
 }
 
@@ -179,6 +183,7 @@
   number={1},
   pages={46--55},
   year={1998},
+  doi={10.1109/99.660313},
   publisher={IEEE}
 }
 
@@ -197,5 +202,6 @@
   number={7981},
   pages={48--52},
   year={2023},
+  doi={10.1038/s41586-023-06551-1},
   publisher={Nature Publishing Group UK London}
 }
