+++ paper.md
@@ -51,7 +51,7 @@ Intrinsic to `G'MIC`'s design are means to map pipelines to commands, advancing 
 `G'MIC` is distributed under the CeCILL free software licenses (GPL-compatible). The core language projects several user interfaces to convert, process or visualize generic *image datasets*. Allied with pipeline toolset, `G'MIC` embodies a highly flexible image model, ranging from 1D scalar signals to 3D+t sequences of multi-spectral volumetric images, hence including 2D color images.
 This makes it a versatile tool for image processing, with a wide range of applications in research, industry and graphic design.
 
-## 1.2. History and Motivation
+## 1.2. Statement of Need
 
 The `G'MIC` project was started in mid-2008 by [David Tschumperlé](https://tschumperle.users.greyc.fr), a research scientist working in the IMAGE team of the _GREYC_, a public research laboratory affiliated with the CNRS institute in France.
 David's area of research is the study and elaboration of image processing algorithms.
@@ -187,312 +187,7 @@ The `G'MIC` source code is available on these various github repositories:
 Last but not least, the project provides regular updates on new developments on social networks such as
 [Mastodon](https://piaille.fr/@gmic) and [Twitter](https://twitter.com/gmic_eu).
 
-## 3. The `G'MIC` Scripting Language
-
-Here we provide some insights about the `G'MIC` scripting language and illustrate its use in the creation of custom image processing algorithms and pipelines.
-
-## 3.1. Overview and Main Features
-
-The `G'MIC` language has been specifically designed for manipulating digital images, in the broadest sense of the term.
-This means, making it easy to apply image processing algorithms that have already been written (e.g. those from the `stdlib`), but also developing new algorithms, either in the form of elaborated pipelines, or in the form of "low-level" operations applied directly to the image pixels.
-This represents a wide variety of different use cases, so some key features of the `G'MIC` language are as follows:
-
-- The ability to handle generic image data type: up to sequences of 4D (3D + channels) hyperspectral images with floating-point valued pixels/voxels.
-- An already implemented large collection of various image processing operators.
-- A powerful math evaluator which supports operations on scalars, complexes, vectors and strings.
-- Multi-threaded computations when possible, using OpenMP [@openmp].
-- Drawing of primitives and 3D rendering capabilities.
-- The ability to manage display windows and handle user events, which leads to the possible creation of simple graphical user interfaces.
-- A network-based update mechanism for the filters and commands.
-
-## 3.2. Language Rules
-
-`G'MIC` is both a _procedural_ and an _interpreted_ language, and therefore shares common properties with equivalent scripting languages
-such as _bash_, _Python_, _LUA_ or _Javascript_, particularly with regard to program structuring and the use of flow control instructions.
-It does, however, have its own distinctive features, some of which we list below:
-
-- **Image storage:**
-At any time, `G'MIC` manages one list of numbered (and optionally named) pixel-based images, entirely stored in computer memory.
-The first image of the list has index `0` and is denoted by `[0]`. The second image of the list is denoted by `[1]`, the third by `[2]` and so on.
[…]
