### 2023-11-13 deb9dc3
Paper length reduction, round #1.
