+++ paper.md
@@ -38,7 +38,7 @@ image retouching applications, such as _GIMP_, _Krita_, _Photoshop_, _Affinity P
 
 Image Analysis, Processing and Filtering, Computer Graphics, Scripting Language, User Interfaces, Creative Coding.
 
-# 1. Introduction
+# 1. Statement of need
 
 ## 1.1. Context
 
@@ -51,7 +51,7 @@ Intrinsic to `G'MIC`'s design are means to map pipelines to commands, advancing 
 `G'MIC` is distributed under the CeCILL free software licenses (GPL-compatible). The core language projects several user interfaces to convert, process or visualize generic *image datasets*. Allied with pipeline toolset, `G'MIC` embodies a highly flexible image model, ranging from 1D scalar signals to 3D+t sequences of multi-spectral volumetric images, hence including 2D color images.
 This makes it a versatile tool for image processing, with a wide range of applications in research, industry and graphic design.
 
-## 1.2. Statement of Need
+## 1.2. History and Motivation
 
 The `G'MIC` project was started in mid-2008 by [David Tschumperlé](https://tschumperle.users.greyc.fr), a research scientist working in the IMAGE team of the _GREYC_, a public research laboratory affiliated with the CNRS institute in France.
 David's area of research is the study and elaboration of image processing algorithms.
@@ -192,28 +192,12 @@ Last but not least, the project provides regular updates on new developments on 
 To illustrate the high degree of genericity of the `G'MIC` framework, we list a selection of a few image processing research projects
 carried out in the [IMAGE team](https://www.greyc.fr/equipes/image) at the [_GREYC_ laboratory](https://www.greyc.fr) (UMR CNRS 6072),
 which have used `G'MIC` for algorithm development, prototyping, testing and result generation.
-The resulting algorithms, presented below, have all been integrated into the `G'MIC` framework, and can now be freely used by users
-(in particular, artists), e.g. via the _G'MIC-Qt_ plug-in.
-This demonstrates the potential of `G'MIC` not only in developing new image processing algorithms,
-but also as a software platform for disseminating these new algorithms to professionals and the general public.
 
 ## 3.1. Patch-Based Image Inpainting
 
-Image _inpainting_ is the name given to the process of replacing part of an image, defined in the form of a user-drawn mask,
-with algorithmically synthesized content.
-This type of algorithms can be used in a number of ways: for example, to remove a distracting element from a photograph (wire, person, vehicle, etc.), or to correct defects in images (scratches, compression noise, etc.).
-The problem of image inpainting has attracted growing interest in recent years, and a number of different algorithmic techniques have been developed
-in an attempt to address the issue.
-Some of these, namely "patch-based" techniques, have proved their effectiveness for the reconstruction of complex patterns, such as textures,
-while maintaining a reasonable algorithmic complexity that allows these algorithms to run on mid-range machines.
-
-Between 2011 and 2015, this problem was studied by researchers Maxime Daisy, Pierre Buyssens, David Tschumperlé, Olivier Le Meur and Olivier Lézoray,
-and this led to the development of new, original image inpainting algorithms, described in detail
+Between 2011 and 2015, the problem of Image _inpainting_ was studied by researchers Maxime Daisy, Pierre Buyssens, David Tschumperlé, Olivier Le Meur and Olivier Lézoray, and this led to the development of new, original image inpainting algorithms, described in detail
 in a publication in _IEEE Transaction on Image Processing_ [@buyssens2015exemplar].
-
-In _G'MIC-Qt_, the _Inpaint [Multi-Scale]_ and _Inpaint [Patch-Based]_ filters implement different patch-based image inpainting algorithms.
[…]
