### 2023-11-13 c6ac496
Paper length reduction, round #2.
