+++ paper.md
@@ -38,17 +38,17 @@ image retouching applications, such as _GIMP_, _Krita_, _Photoshop_, _Affinity P
 
 Image Analysis, Processing and Filtering, Computer Graphics, Scripting Language, User Interfaces, Creative Coding.
 
-# 1. Statement of need
+# 1. Statement of Need
 
 ## 1.1. Context
 
-Tools not only shape work but set limits as to how far such work can go. Tools with limited capabilities or rigid behaviors not only circumscribe results but also constrain tool-users' awareness of what is possibile — they suppose from such inflexible tools only notions of what cannot be done.
+Intrinsic to `G'MIC`'s design are means to map image processing pipelines to commands, advancing the tool as a self-extending language
+and fortifying how users' conduct their work. Primal command pipelines may be further assembled into those having wider remits, these suitably
+named to bespeak their extended purposes and available for succeeding command prototyping.
 
-Intrinsic to `G'MIC`'s design are means to map pipelines to commands, advancing the tool as a self-extending language and fortifying how users' conduct their work. Primal command pipelines may be further assembled into those having wider remits, these suitably named to bespeak their extended purposes and available for succeeding command prototyping.
-
-`G'MIC` itself is based upon primitives of the _C`++`_ library [`CImg`](http://cimg.eu) [@cimg] that are broadly applicable to image processing work and which have been optimized for CPU performance. Most commands in the standard `G'MIC` distribution extend these primitives, using the aforementioned scheme. It is one which remains available to users, extending the language in line with their specific domains of expertise and making the language "their own".
-
-`G'MIC` is distributed under the CeCILL free software licenses (GPL-compatible). The core language projects several user interfaces to convert, process or visualize generic *image datasets*. Allied with pipeline toolset, `G'MIC` embodies a highly flexible image model, ranging from 1D scalar signals to 3D+t sequences of multi-spectral volumetric images, hence including 2D color images.
+`G'MIC` is distributed under the CeCILL free software licenses (GPL-compatible). The core language projects several user interfaces to convert,
+process or visualize generic *image datasets*. Allied with pipeline toolset, `G'MIC` embodies a highly flexible image model,
+ranging from 1D scalar signals to 3D+t sequences of multi-spectral volumetric images, hence including 2D color images.
 This makes it a versatile tool for image processing, with a wide range of applications in research, industry and graphic design.
 
 ## 1.2. History and Motivation
@@ -57,8 +57,6 @@ The `G'MIC` project was started in mid-2008 by [David Tschumperlé](https://tsch
 David's area of research is the study and elaboration of image processing algorithms.
 
 To that end, he first began developing [`CImg`](http://cimg.eu) [@cimg], beginning in 1999 and continuing to the present. `CImg` is  an open-source _C`++`_ library for generic image processing. Here, _generic_ implies a library that addresses structurally diverse imagery: photographs, multi-spectral images (e.g. from satellites), medical images (MRI, X-ray, tomography, etc.) and technical animations, among others.
-The `CImg` library has therefore been designed to handle a wide variety of different image types, whether defined on 2D or 3D grids, or with any number of channels.
-He still develops and maintains this library.
 
 That said, `CImg` exhibits certain limitations for everyday research work:
 
@@ -72,13 +70,7 @@ These limitations motivated `G'MIC`'s development, beginning in 2008. Two design
 
[…]
