+++ paper.bib
@@ -109,7 +109,7 @@
 
 @misc{jehan18,
   title={{L}ineart {B}ucket {F}ill},
-  author={GIMP Developers},
+  author={GIMP},
   url={https://developer.gimp.org/core/algorithm/line-art-bucket-fill/},
   year={2018}
 }
