### 2023-11-13 ab77c01
Paper length reduction, round #3.
