+++ paper.md
@@ -30,7 +30,7 @@ bibliography: paper.bib
 
 We present `G'MIC`, an open-source self-extending framework that defines an original, concise, scripting language for the writing of possibly complex
 image processing operators and pipelines.
-`G'MIC` also provides several user interfaces allowing for the manipulation of digital images, adapted to different levels of user expertise,
+`G'MIC` provides several user interfaces allowing for the manipulation of digital images, adapted to different levels of user expertise,
 either from the command line, or as a _C/C`++`_ library, or as a user-friendly graphical plug-in that extends the capabilities of popular digital
 image retouching applications, such as _GIMP_, _Krita_, _Photoshop_, _Affinity Photo_ and others.
 
@@ -42,9 +42,8 @@ Image Analysis, Processing and Filtering, Computer Graphics, Scripting Language,
 
 ## 1.1. Context
 
-Intrinsic to `G'MIC`'s design are means to map image processing pipelines to commands, advancing the tool as a self-extending language
-and fortifying how users' conduct their work. Primal command pipelines may be further assembled into those having wider remits, these suitably
-named to bespeak their extended purposes and available for succeeding command prototyping.
+Intrinsic to `G'MIC`'s design are means to map image processing pipelines to commands, advancing the tool as a self-extending language.
+Primal command pipelines may be further assembled into those having wider remits, these suitably named to bespeak their extended purposes and available for succeeding command prototyping.
 
 `G'MIC` is distributed under the CeCILL free software licenses (GPL-compatible). The core language projects several user interfaces to convert,
 process or visualize generic *image datasets*. Allied with pipeline toolset, `G'MIC` embodies a highly flexible image model,
@@ -53,16 +52,17 @@ This makes it a versatile tool for image processing, with a wide range of applic
 
 ## 1.2. History and Motivation
 
-The `G'MIC` project was started in mid-2008 by [David Tschumperlé](https://tschumperle.users.greyc.fr), a research scientist working in the IMAGE team of the _GREYC_, a public research laboratory affiliated with the CNRS institute in France.
-David's area of research is the study and elaboration of image processing algorithms.
+The `G'MIC` project was started in mid-2008 by research scientists working in the IMAGE team of the _GREYC_ (public research laboratory in France),.
+whose area of research is the elaboration of image processing algorithms.
 
-To that end, he first began developing [`CImg`](http://cimg.eu) [@cimg], beginning in 1999 and continuing to the present. `CImg` is  an open-source _C`++`_ library for generic image processing. Here, _generic_ implies a library that addresses structurally diverse imagery: photographs, multi-spectral images (e.g. from satellites), medical images (MRI, X-ray, tomography, etc.) and technical animations, among others.
+To that end, they first began developing [`CImg`](http://cimg.eu) [@cimg], beginning in 1999 and continuing to the present.
+`CImg` is an open-source _C`++`_ library for generic image processing. Here, _generic_ implies a library that addresses structurally diverse imagery: photographs, multi-spectral images (e.g. from satellites), medical images (MRI, X-ray, tomography, etc.) and technical animations, among others.
 
 That said, `CImg` exhibits certain limitations for everyday research work:
 
-1. When one simply wants to apply a predefined algorithm from `CImg` to an image, one needs to write a small, _C`++`_ program. Perhaps it is only a few lines long, but still it must be compiled and linked — and possibly debugged — before it can be executed. In the context of research work, such mechanics are just so many distractions. The idea of being able to run those algorithms directly from the command line is tempting.
[…]
