### 2023-11-13 fe685cc
Paper length reduction, round #4.
