+++ paper.md
@@ -2,11 +2,9 @@
 title: "`G'MIC`: An Open-Source Self-Extending Framework for Image Processing"
 tags:
   - image processing
-  - image filtering
   - computer graphics
   - scripting language
   - user interfaces
-  - creative coding
 authors:
   - name: David Tschumperlé
     corresponding: true
@@ -20,7 +18,7 @@ authors:
 affiliations:
  - name: GREYC Lab (IMAGE Team), CNRS, Normandie Univ, UNICAEN, ENSICAEN, F-14000 Caen, France 
    index: 1
- - name: Independent contributor, New York City, US
+ - name: Independent researcher, New York City, US
    index: 2
 date: 8 June 2023
 bibliography: paper.bib
@@ -36,14 +34,15 @@ image retouching applications, such as _GIMP_, _Krita_, _Photoshop_, _Affinity P
 
 ## Keywords
 
-Image Analysis, Processing and Filtering, Computer Graphics, Scripting Language, User Interfaces, Creative Coding.
+Image Analysis and Processing, Computer Graphics, Scripting Language, User Interfaces.
 
 # 1. Statement of Need
 
 ## 1.1. Context
 
 Intrinsic to `G'MIC`'s design are means to map image processing pipelines to commands, advancing the tool as a self-extending language.
-Primal command pipelines may be further assembled into those having wider remits, these suitably named to bespeak their extended purposes and available for succeeding command prototyping.
+Primal command pipelines may be further assembled into those having wider remits, these suitably named to bespeak their extended purposes
+and available for succeeding command prototyping.
 
 `G'MIC` is distributed under the CeCILL free software licenses (GPL-compatible). The core language projects several user interfaces to convert,
 process or visualize generic *image datasets*. Allied with pipeline toolset, `G'MIC` embodies a highly flexible image model,
@@ -94,13 +93,18 @@ While these software somehow allows the user to create its own pipeline of image
 
 ## 2.1. Core Components
 
-The current architecture of the `G'MIC` framework is depicted on Fig. 1. This corresponds to the current state of the framework (version **3.3.2**), at the time of writing.
+The current architecture of the `G'MIC` framework is depicted on Fig. 1.
 
 ![Overview of the `G'MIC` framework.](images/gmic_architecture.png)
 
-The organization of this framework revolves around a central component: the **`G'MIC` scripting language interpreter** (in yellow). This interpreter uses the native functionalities of the **`CImg` library** (which is implemented in _C`++`_, in blue), but relies also on a set of commands, written in the `G'MIC` language themselves, constituting a **_standard library_ (`stdlib`)** for the framework (in green). The other components (in orange) stand for the different user interfaces provided by the framework. More than 1000 distinct commands are currently implemented in the `stdlib`, covering a large portion of general image processing needs.
+The framework revolves around a central component: the **`G'MIC` scripting language interpreter** (yellow).
+This interpreter uses the native functionalities of the **`CImg` library** (implemented in _C`++`_, blue),
+but relies also on a set of commands, written in the `G'MIC` language themselves, constituting a **_standard library_ (`stdlib`)**
+for the framework (green). The other components (orange) stand for the different user interfaces provided by the framework.
+More than 1000 distinct commands are currently implemented, covering a large portion of image processing needs.
 
-The `G'MIC` interpreter lets the user write and run custom programs using this predefined set of commands, for tasks as varied as writing new image filters, implementing generative algorithms or creating user interfaces for image manipulation.
[…]
