### 2023-11-13 003086d
Paper length reduction, round #5.
