+++ paper.md
@@ -2,7 +2,6 @@
 title: "`G'MIC`: An Open-Source Self-Extending Framework for Image Processing"
 tags:
   - image processing
-  - computer graphics
   - scripting language
   - user interfaces
 authors:
@@ -20,13 +19,13 @@ affiliations:
    index: 1
  - name: Independent researcher, New York City, US
    index: 2
-date: 8 June 2023
+date: 13 november 2023
 bibliography: paper.bib
 ---
 
 ## Abstract
 
-We present `G'MIC`, an open-source self-extending framework that defines an original, concise, scripting language for the writing of possibly complex
+We present `G'MIC`, an open-source self-extending framework that defines an original, concise, scripting language for writing possibly complex
 image processing operators and pipelines.
 `G'MIC` provides several user interfaces allowing for the manipulation of digital images, adapted to different levels of user expertise,
 either from the command line, or as a _C/C`++`_ library, or as a user-friendly graphical plug-in that extends the capabilities of popular digital
@@ -34,7 +33,7 @@ image retouching applications, such as _GIMP_, _Krita_, _Photoshop_, _Affinity P
 
 ## Keywords
 
-Image Analysis and Processing, Computer Graphics, Scripting Language, User Interfaces.
+Image Analysis and Processing, Scripting Language, User Interfaces.
 
 # 1. Statement of Need
 
@@ -44,90 +43,92 @@ Intrinsic to `G'MIC`'s design are means to map image processing pipelines to com
 Primal command pipelines may be further assembled into those having wider remits, these suitably named to bespeak their extended purposes
 and available for succeeding command prototyping.
 
-`G'MIC` is distributed under the CeCILL free software licenses (GPL-compatible). The core language projects several user interfaces to convert,
-process or visualize generic *image datasets*. Allied with pipeline toolset, `G'MIC` embodies a highly flexible image model,
-ranging from 1D scalar signals to 3D+t sequences of multi-spectral volumetric images, hence including 2D color images.
+`G'MIC` is distributed under the CeCILL licenses. The core language projects several user interfaces to convert,
+process or visualize *image datasets*. Allied with pipeline toolset, `G'MIC` embodies a highly flexible image model,
+ranging from 1D signals to 3D+t sequences of multi-spectral volumetric images, hence including 2D color images.
 This makes it a versatile tool for image processing, with a wide range of applications in research, industry and graphic design.
 
 ## 1.2. History and Motivation
 
-The `G'MIC` project was started in mid-2008 by research scientists working in the IMAGE team of the _GREYC_ (public research laboratory in France),.
-whose area of research is the elaboration of image processing algorithms.
+The `G'MIC` project was initiated in 2008 by research scientists of the IMAGE team at the _GREYC_ laboratory, a public research lab in France.
+Their area of research focuses on development of image processing algorithms.
 
 To that end, they first began developing [`CImg`](http://cimg.eu) [@cimg], beginning in 1999 and continuing to the present.
-`CImg` is an open-source _C`++`_ library for generic image processing. Here, _generic_ implies a library that addresses structurally diverse imagery: photographs, multi-spectral images (e.g. from satellites), medical images (MRI, X-ray, tomography, etc.) and technical animations, among others.
+`CImg` is an open-source _C`++`_ library for generic image processing, which means a library that is able to address structurally diverse imagery:
+photographs, multi-spectral images, medical images (MRI, X-ray, tomography, etc.) and animations, among others.
 
 That said, `CImg` exhibits certain limitations for everyday research work:
 
[…]
