### 2023-11-13 #6056 @danielskatz
The issue with the statement of need section might be because you are adding numbers to your section headings, which the JOSS example paper doesn't have. Can you remove them and check again?
