### 2023-11-13 #6056 @danielskatz
@dtschump - that seems to have fixed it.  I'll next look for an editor.  While I do so, if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission, or suggest people from outside JOSS.
