### 2023-11-13 #6056 @dtschump
About reviewers suggestion :

In your list, I think people with an image processing & C/C++ background, would be good candidates.
So with the search filters, I found :  [Jonathan Bach Martin](https://reviewers.joss.theoj.org/reviewers/3141), [Justus Schock](https://reviewers.joss.theoj.org/reviewers/1261) and [@hechth](https://reviewers.joss.theoj.org/reviewers/1720).

Outside JOSS official reviewers, I would recommend :
-  **Pierre Kornprobst** (https://team.inria.fr/biovision/pierre-kornprobst/), as an experienced researcher who has a long experience in implementing image processing algorithms in C/C++, and supervising students who do so, for various research projects.
- **Hughes Talbot** (https://hugues-talbot.github.io/), who participated in the development of the [PINK Image processing library](http://www.ipol.im/event/2012_imlib/docs/pink.pdf) (opensource, C++, kind of "concurrent" to G'MIC with a quite similar spirit).
- **Alexandre Duretz-Lutz** (https://www.lrde.epita.fr/~adl/) or Thierry Géraud (https://www.lrde.epita.fr/wiki/User:Theo), who both initiated the development of the [OLENA image processing library](https://www.lrde.epita.fr/wiki/Olena/Download) (also opensource, C++, used for internal research in the EPITA research lab).
