### 2023-11-15 d3f91c2
Fix wrong parsing of 'xN' repetitions in 'input' command.
