### 2023-11-24 0ecf087
Make 'gmic_stdlib_community' default for all G'MIC interfaces.
