+++ src/Makefile
@@ -621,7 +621,7 @@ gmic_qt_shared: gmic_qt_files CImg.h gmic_stdlib_community.h
 
 # ZArt.
 #------
-zart_shared: zart_files CImg.h gmic_stdlib.h
+zart_shared: zart_files CImg.h gmic_stdlib_community.h
 ifeq ($(OS),Darwin)
 	cd ../zart && $(QMAKE) CONFIG+=release GMIC_PATH=$(SRC_PATH) GMIC_DYNAMIC_LINKING=on zart.pro && $(MAKE) && $(STRIP) zart.app/Contents/MacOS/zart
 else
@@ -629,7 +629,7 @@ else
 endif
 	@echo "Executable 'zart' has been successfully compiled in '../zart/'."
 
-zart: zart_files CImg.h gmic_stdlib.h
+zart: zart_files CImg.h gmic_stdlib_community.h
 ifeq ($(OS),Darwin)
 	cd ../zart && $(QMAKE) CONFIG+=release GMIC_PATH=$(SRC_PATH) zart.pro && $(MAKE) && $(STRIP) zart.app/Contents/MacOS/zart
 else
@@ -671,74 +671,6 @@ CImg.h:
           touch CImg.h; \
         fi
 	@echo " done!"
-
-gmic_stdlib.h: gmic_stdlib.gmic
-	@echo "> Retrieve G'MIC Standard Library..."
-	@if [ ! -f ./gmic_stdlib.h ]; then \
-	  $(WGET) gmic_stdlib.h https://gmic.eu/gmic_stdlib$(VERSION).h; \
-	  touch gmic_stdlib.h; \
-	fi
-	@echo " done!"
-
-do_gmic_stdlib.h:
-	@echo "/*\n\
- #\n\
- #  File        : gmic_stdlib.h (v.$(SVERSION))\n\
- #                ( C++ header file )\n\
- #\n\
- #  Description : Raw data arrays encoding the G'MIC standard library\n\
- #                into a compressed form.\n\
- #                This file has been automatically generated by the Makefile entry\n\
- #                'do_gmic_stdlib.h:', from the G'MIC source file 'gmic_stdlib.gmic'.\n\
- #                ( https://gmic.eu )\n\
- #\n\
- #  Copyright   : David Tschumperlé\n\
- #                ( https://tschumperle.users.greyc.fr/ )\n\
- #\n\
- #  Licenses    : This file is 'dual-licensed', you have to choose one\n\
- #      	   of the two licenses below to apply.\n\
- #\n\
- #                CeCILL-C\n\
- #                The CeCILL-C license is close to the GNU LGPL.\n\
- #                ( http://cecill.info/licences/Licence_CeCILL-C_V1-en.html )\n\
- #\n\
- #            or  CeCILL v2.1\n\
- #                The CeCILL license is compatible with the GNU GPL.\n\
- #                ( http://cecill.info/licences/Licence_CeCILL_V2.1-en.html )\n\
- #\n\
- #  This software is governed either by the CeCILL or the CeCILL-C license\n\
- #  under French law and abiding by the rules of distribution of free software.\n\
- #  You can  use, modify and or redistribute the software under the terms of\n\
- #  the CeCILL or CeCILL-C licenses as circulated by CEA, CNRS and INRIA\n\
- #  at the following URL: \"http://cecill.info\".\n\
- #\n\
- #  As a counterpart to the access to the source code and  rights to copy,\n\
- #  modify and redistribute granted by the license, users are provided only\n\
- #  with a limited warranty  and the software's author,  the holder of the\n\
- #  economic rights,  and the successive licensors  have only  limited\n\
- #  liability.\n\
- #\n\
[…]
