### 2023-11-24 bddb87a
Make 'gmic_stdlib_community' default for all G'MIC interfaces.
