### 2023-11-27 31bdc5a
Interpreter now discards possible comments passed in the 'commands_line' parameter of functions in the G'MIC API.
