### 2023-12-30 b5e5014
cimg_math_parser(): New function 'index()'.
