### 2023-12-30 a452ff3
cimg_math_parser(): New function 'noise()'.
