### 2024-01-03 ee24a78
Command 'rand' now accepts a '[pdf]' parameter that sets the probability density function followed by the random values.
