+++ src/gmic.cpp
@@ -11050,11 +11050,15 @@ gmic& gmic::_run(const CImgList<char>& commands_line, unsigned int& position,
         // Fill with random values.
         if (!std::strcmp("rand",command)) {
           gmic_substitute_args(true);
-          ind0.assign(); ind1.assign();
-          sep0 = sep1 = *argx = *argy = *indices = 0;
+          ind.assign(); ind0.assign(); ind1.assign();
+          sep = sep0 = sep1 = *argx = *argy = *argz = *indices = 0;
           value0 = value1 = 0;
-          if (cimg_sscanf(argument,"%255[][a-zA-Z0-9_.eE%+-],%255[][a-zA-Z0-9_.eE%+-]%c",
-                          gmic_use_argx,gmic_use_argy,&end)==2 &&
+          if ((cimg_sscanf(argument,"%255[][a-zA-Z0-9_.eE%+-],%255[][a-zA-Z0-9_.eE%+-]%c",
+                           gmic_use_argx,gmic_use_argy,&end)==2 ||
+               (cimg_sscanf(argument,"%255[][a-zA-Z0-9_.eE%+-],%255[][a-zA-Z0-9_.eE%+-],[%255[a-zA-Z0-9_.%+-]%c%c",
+                            argx,argy,gmic_use_argz,&sep,&end)==4 &&
+                sep==']' &&
+                (ind=selection2cimg(argz,images.size(),images_names,"rand")).height()==1)) &&
               ((cimg_sscanf(argx,"[%255[a-zA-Z0-9_.%+-]%c%c",gmic_use_indices,&sep0,&end)==2 &&
                 sep0==']' &&
                 (ind0=selection2cimg(indices,images.size(),images_names,"rand")).height()==1) ||
@@ -11067,10 +11071,17 @@ gmic& gmic::_run(const CImgList<char>& commands_line, unsigned int& position,
                cimg_sscanf(argy,"%lf%c",&value1,&end)==1)) {
             if (ind0) { value0 = images[*ind0].min(); sep0 = 0; }
             if (ind1) { value1 = images[*ind1].max(); sep1 = 0; }
-            print(0,"Fill image%s with random values, in range [%g%s,%g%s].",
-                  gmic_selection.data(),
-                  value0,sep0=='%'?"%":"",
-                  value1,sep1=='%'?"%":"");
+            if (ind)
+              print(0,"Fill image%s with random values in range [%g%s,%g%s] (with distribution [%u]).",
+                    gmic_selection.data(),
+                    value0,sep0=='%'?"%":"",
+                    value1,sep1=='%'?"%":"",
+                    *ind);
+            else
+              print(0,"Fill image%s with random values in range [%g%s,%g%s] (uniformly distributed).",
+                    gmic_selection.data(),
+                    value0,sep0=='%'?"%":"",
+                    value1,sep1=='%'?"%":"");
             cimg_forY(selection,l) {
               CImg<T>& img = gmic_check(images[selection[l]]);
               nvalue0 = value0; nvalue1 = value1;
@@ -11080,7 +11091,8 @@ gmic& gmic::_run(const CImgList<char>& commands_line, unsigned int& position,
                 if (sep0=='%') nvalue0 = vmin + (vmax - vmin)*value0/100;
                 if (sep1=='%') nvalue1 = vmin + (vmax - vmin)*value1/100;
               }
-              gmic_apply(rand((T)nvalue0,(T)nvalue1),true);
+              if (ind) { gmic_apply(rand((T)nvalue0,(T)nvalue1,images[*ind]),true); }
+              else { gmic_apply(rand((T)nvalue0,(T)nvalue1),true); }
             }
[…]
