### 2024-01-11 #6056 @jamesrhester
Happy to review this, bearing in mind that I'm not an image processing expert, have just messed around a bit with image processing tools.
