### 2024-01-12 #6056 @gkthiruvathukal
@jamesrhester It isn't required to have deep expertise to review JOSS submissions.  Having prior experience with image processing tools is a plus. The goal is to evaluate that the software does what the authors say it does and meets the JOSS requirements. I'm going to go ahead and assign you and see if I can find a second reviewer in the meantime.
