+++ src/gmic.cpp
@@ -5014,7 +5014,7 @@ gmic& gmic::_run(const CImgList<char>& commands_line, unsigned int& position,
   CImgList<unsigned int> primitives;
   CImgList<unsigned char> g_list_uc;
   CImgList<float> g_list_f;
-  CImgList<char> g_list_c, arguments;
+  CImgList<char> g_list_c, g_arguments;
   CImgList<T> g_list;
 
   CImg<unsigned int> ind, ind0, ind1;
@@ -13422,8 +13422,8 @@ gmic& gmic::_run(const CImgList<char>& commands_line, unsigned int& position,
             }
 
             // Set $0 to be the command name.
-            arguments.assign(1);
-            CImg<char>::string(command).move_to(arguments[0]);
+            if (g_arguments.is_empty()) g_arguments.assign(1); else cimglist_for(g_arguments,l) g_arguments[l].assign();
+            CImg<char>::string(command).move_to(g_arguments[0]);
             unsigned int nb_arguments = 0;
 
             if (commands_has_arguments[hash_custom](ind_custom,0)) { // Command takes arguments
@@ -13434,15 +13434,15 @@ gmic& gmic::_run(const CImgList<char>& commands_line, unsigned int& position,
                 if ((_ss=std::strchr(ss,','))!=0) {
                   if (ss==_ss) ++nb_arguments;
                   else {
-                    if (++nb_arguments>=arguments.size()) arguments.insert(nb_arguments + 1 - arguments.size());
+                    if (++nb_arguments>=g_arguments.size()) g_arguments.insert(nb_arguments + 1 - g_arguments.size());
                     CImg<char> arg_item(ss,(unsigned int)(_ss - ss + 1));
                     arg_item.back() = 0;
-                    arg_item.move_to(arguments[nb_arguments]);
+                    arg_item.move_to(g_arguments[nb_arguments]);
                   }
                 } else {
                   if (*ss) {
-                    if (++nb_arguments>=arguments.size()) arguments.insert(nb_arguments + 1 - arguments.size());
-                    if (*ss!=',') CImg<char>::string(ss).move_to(arguments[nb_arguments]);
+                    if (++nb_arguments>=g_arguments.size()) g_arguments.insert(nb_arguments + 1 - g_arguments.size());
+                    if (*ss!=',') CImg<char>::string(ss).move_to(g_arguments[nb_arguments]);
                   }
                   break;
                 }
@@ -13452,7 +13452,7 @@ gmic& gmic::_run(const CImgList<char>& commands_line, unsigned int& position,
                       nb_arguments,nb_arguments!=1?"s":"",
                       command_name,nb_arguments>0?":":".");
                 for (unsigned int i = 1; i<=nb_arguments; ++i)
-                  if (arguments[i]) debug("  $%d = '%s'",i,arguments[i].data());
+                  if (g_arguments[i]) debug("  $%d = '%s'",i,g_arguments[i].data());
                   else debug("  $%d = (undefined)",i);
               }
             }
@@ -13509,7 +13509,7 @@ gmic& gmic::_run(const CImgList<char>& commands_line, unsigned int& position,
                   nsource+=4;
                   for (unsigned int i = 1; i<=nb_arguments; ++i) {
                     CImg<char>::append_string_to('\"',substituted_command,ptr_sub);
-                    CImg<char>(arguments[i].data(),arguments[i].width() - 1,1,1,1,true).
+                    CImg<char>(g_arguments[i].data(),g_arguments[i].width() - 1,1,1,1,true).
                       append_string_to(substituted_command,ptr_sub);
                     CImg<char>::append_string_to('\"',substituted_command,ptr_sub);
[…]
