+++ src/gmic.cpp
@@ -5014,7 +5014,7 @@ gmic& gmic::_run(const CImgList<char>& commands_line, unsigned int& position,
   CImgList<unsigned int> primitives;
   CImgList<unsigned char> g_list_uc;
   CImgList<float> g_list_f;
-  CImgList<char> g_list_c, g_arguments;
+  CImgList<char> g_list_c;
   CImgList<T> g_list;
 
   CImg<unsigned int> ind, ind0, ind1;
@@ -10322,19 +10322,19 @@ gmic& gmic::_run(const CImgList<char>& commands_line, unsigned int& position,
           }
           if (!*_arg) print(0,"Skip command 'parallel' (no commands specified)."); // No command specified
           else {
-            CImgList<char> l_arguments = CImg<char>::string(_arg).get_split(CImg<char>::vector(','),0,false);
-            CImg<_gmic_parallel<T> >(1,l_arguments.width()).move_to(gmic_threads);
+            CImg<char>::string(_arg).get_split(CImg<char>::vector(','),0,false).move_to(g_list_c);
+            CImg<_gmic_parallel<T> >(1,g_list_c.width()).move_to(gmic_threads);
             CImg<_gmic_parallel<T> > &_gmic_threads = gmic_threads.back();
 
 #ifdef gmic_is_parallel
             print(0,"Execute %d command%s '%s' in parallel%s.",
-                  l_arguments.width(),l_arguments.width()>1?"s":"",_arg_text,
+                  g_list_c.width(),g_list_c.width()>1?"s":"",_arg_text,
                   wait_mode?" and wait thread termination immediately":
                   " and wait thread termination when current environment ends");
 #else // #ifdef gmic_is_parallel
             print(0,"Execute %d commands '%s' (run sequentially, "
                   "parallel computing disabled).",
-                  l_arguments.width(),_arg_text);
+                  g_list_c.width(),_arg_text);
 #endif // #ifdef gmic_is_parallel
 
             // Prepare thread structures.
@@ -10359,15 +10359,16 @@ gmic& gmic::_run(const CImgList<char>& commands_line, unsigned int& position,
               _gmic_threads[l].is_thread_running = true;
 
               // Substitute special characters codes appearing outside strings.
-              l_arguments[l].resize(1,l_arguments[l].height() + 1,1,1,0);
+              g_list_c[l].resize(1,g_list_c[l].height() + 1,1,1,0);
               bool is_dquoted = false;
-              for (char *s = l_arguments[l].data(); *s; ++s) {
+              for (char *s = g_list_c[l].data(); *s; ++s) {
                 if (*s=='\"') is_dquoted = !is_dquoted;
                 else if (!is_dquoted) _strreplace_fw(*s);
               }
-              gmic_instance.commands_line_to_CImgList(l_arguments[l].data()).
+              gmic_instance.commands_line_to_CImgList(g_list_c[l].data()).
                 move_to(_gmic_threads[l].commands_line);
             }
+            g_list_c.assign();
 
             // Run threads.
             cimg_forY(_gmic_threads,l) {
@@ -13422,8 +13423,8 @@ gmic& gmic::_run(const CImgList<char>& commands_line, unsigned int& position,
             }
 
             // Set $0 to be the command name.
-            if (g_arguments.is_empty()) g_arguments.assign(1); else cimglist_for(g_arguments,l) g_arguments[l].assign();
-            CImg<char>::string(command).move_to(g_arguments[0]);
+            g_list_c.assign(1); // Used to store arguments.
+            CImg<char>::string(command).move_to(g_list_c[0]);
             unsigned int nb_arguments = 0;
 
             if (commands_has_arguments[hash_custom](ind_custom,0)) { // Command takes arguments
[…]
