+++ src/gmic.cpp
@@ -7866,7 +7866,7 @@ gmic& gmic::_run(const CImgList<char>& commands_line, unsigned int& position,
             if (resolution<1) resolution = 65536;
 
             gmic_use_argx;
-            cimg_snprintf(argx,_argx.width(),"x = lerp(%g,%g,x/%d);",
+            cimg_snprintf(argx,_argx.width(),"x = lerp(%.17g,%.17g,x/%d);",
                           xmin,xmax,(unsigned int)(resolution>1?resolution - 1:0));
             const CImg<char> n_formula = (CImg<char>::string(argx,false,true),
                                           CImg<char>::string(formula,true,true))>'x';
@@ -10648,7 +10648,7 @@ gmic& gmic::_run(const CImgList<char>& commands_line, unsigned int& position,
             if (resolution<1) resolution = 65536;
 
             gmic_use_argx;
-            cimg_snprintf(argx,_argx.width(),"x = lerp(%g,%g,x/%d);",
+            cimg_snprintf(argx,_argx.width(),"x = lerp(%.17g,%.17g,x/%d);",
                           xmin,xmax,(unsigned int)(resolution>1?resolution - 1:0));
             const CImg<char> n_formula = (CImg<char>::string(argx,false,true),
                                           CImg<char>::string(formula,true,true))>'x';
