### 2024-02-26 fb1efab
Command 'check': skip argument checking if called in sub-command.
