### 2024-02-29 #6056 @drj11
i can review. I'm not exactly expert, but i have written the pure-Python PNG module PyPNG https://gitlab.com/drj11/pypng
