### 2024-04-13 #6618 @editorialbot
[REVIEW]: G'MIC: An Open-Source Self-Extending Framework for Image Processing
**Submitting author:** @dtschump (<a href="http://orcid.org/0000-0003-3454-5079">David Tschumperlé</a>)
**Repository:** https://github.com/GreycLab/gmic
**Branch with paper.md** (empty if default branch): joss
**Version:** 3.4.3
