### 2024-05-28 #6618 @Smattr
Like jamesrhester, I just haven’t had time yet. My previous comment about not being available until May was not a request to be assigned as a reviewer, assuming I would immediately start May 1 :wink:. I think I should have some time upcoming, but fair warning I have not taken even a cursory glance at the paper yet to see if I’m qualified to review.
