### 2024-06-08 GreycLab/gmic#53 @Smattr
[JOSS] ambiguous library description
The JOSS paper being reviewed at https://github.com/openjournals/joss-reviews/issues/6618 contains the sentence “libgmic and libgmic, respectively C++ and C libraries…” Presumably this is not meant to say the identical text “libgmic” twice?
