### 2024-06-09 GreycLab/gmic#54 @Smattr
[JOSS] community guidelines
One of the checklist items for the JOSS review happening at https://github.com/openjournals/joss-reviews/issues/6618 is:

> **Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I can’t immediately see something that fits this. Maybe the repository should gain a CONTRIBUTING.md? If I’ve missed something existing, please let me know.
