### 2024-06-09 GreycLab/gmic#55 @Smattr
[JOSS] project does not build
(In relation to JOSS review at https://github.com/openjournals/joss-reviews/issues/6618)

On the latest commit (f2d22d64f3bfd84e37b21b18d23414ecc8ae16b9), after downloading https://gmic.eu/gmic_stdlib_community337.h (SHA256 601e267b08907773c2a6d798331d22092c38d15a4f839ca80ea359bd413898cf at time of writing), configure and build does not seem to work:

```bash
$ cmake -B build -S .
-- The CXX compiler identification is GNU 11.4.0
-- The C compiler identification is GNU 11.4.0
-- Detecting CXX compiler ABI info
-- Check for working CXX compiler: /usr/bin/c++ - skipped
-- Detecting CXX compile features
-- Detecting C compiler ABI info
-- Detecting C compiler ABI info - done
-- Check for working C compiler: /usr/bin/cc - skipped
-- Detecting C compile features
-- Detecting C compile features - done
-- Could NOT find PkgConfig (missing: PKG_CONFIG_EXECUTABLE)
-- Found CImg: /tmp/tmp.W1ovy1i20l/gmic/src/CImg.h
-- Found OpenMP_C: -fopenmp (found version "4.5")
-- Found OpenMP_CXX: -fopenmp (found version "4.5")
-- Found OpenMP: TRUE (found version "4.5")
-- Found ZLIB: /usr/lib/x86_64-linux-gnu/libz.so (found version "1.2.11")
-- Could NOT find CURL (missing: CURL_LIBRARY CURL_INCLUDE_DIR)
-- Could NOT find X11 (missing: X11_X11_INCLUDE_PATH X11_X11_LIB)
-- Could NOT find PkgConfig (missing: PKG_CONFIG_EXECUTABLE)
-- Could NOT find Fftw (missing: FFTW3_LIBRARIES)
-- Could NOT find PkgConfig (missing: PKG_CONFIG_EXECUTABLE)
-- Could NOT find GraphicsMagick (missing: GRAPHICSMAGICK_LIBRARIES)
-- Could NOT find TIFF (missing: TIFF_LIBRARY TIFF_INCLUDE_DIR)
-- Could NOT find PNG (missing: PNG_LIBRARY PNG_PNG_INCLUDE_DIR)
-- Could NOT find JPEG (missing: JPEG_LIBRARY JPEG_INCLUDE_DIR)
-- Could NOT find PkgConfig (missing: PKG_CONFIG_EXECUTABLE)
-- Could NOT find OpenEXR (missing: OPENEXR_LIBRARIES)
-- Looking for pthread.h
-- Looking for pthread.h - found
-- Performing Test CMAKE_HAVE_LIBC_PTHREAD
-- Performing Test CMAKE_HAVE_LIBC_PTHREAD - Success
-- Found Threads: TRUE
-- Found GMicStdlib: /tmp/tmp.W1ovy1i20l/gmic/src/gmic_stdlib_community.h
-- The following OPTIONAL packages have been found:

 * OpenMP
 * ZLIB
 * Threads

-- The following REQUIRED packages have been found:

 * CImg
 * GMicStdlib

-- The following OPTIONAL packages have not been found:

 * CURL
 * X11
 * Fftw
 * GraphicsMagick
 * TIFF
 * PNG
 * JPEG
 * PkgConfig
 * OpenEXR

-- Configuring done
-- Generating done
-- Build files have been written to: /tmp/tmp.W1ovy1i20l/gmic/build
$ cmake --build build
[ 12%] Building CXX object CMakeFiles/libgmic.dir/src/gmic.cpp.o
/tmp/tmp.W1ovy1i20l/gmic/src/gmic.cpp:1788:1: error: expected \u2018}\u2019 before \u2018using\u2019
 1788 | using namespace gmic_library;
      | ^~~~~
In file included from /tmp/tmp.W1ovy1i20l/gmic/src/gmic.cpp:1787:
/tmp/tmp.W1ovy1i20l/gmic/src/gmic_stdlib_community.h:55:35: note: to match this \u2018{\u2019
   55 | const unsigned char data_gmic[] = {
      |                                   ^
/tmp/tmp.W1ovy1i20l/gmic/src/gmic.cpp:1788:1: error: expected \u2018,\u2019 or \u2018;\u2019 before \u2018using\u2019
 1788 | using namespace gmic_library;
      | ^~~~~
/tmp/tmp.W1ovy1i20l/gmic/src/gmic.cpp: In static member function \u2018static const gmic_library::gmic_image<char>& gmic::decompress_stdlib()\u2019:
/tmp/tmp.W1ovy1i20l/gmic/src/gmic.cpp:2741:71: error: \u2018size_data_gmic\u2019 was not declared in this scope; did you mean \u2018data_gmic\u2019?
 2741 |       CImgList<char>::get_unserialize(CImg<unsigned char>(data_gmic,1,size_data_gmic,1,1,true))[0].
      |                                                                       ^~~~~~~~~~~~~~
      |                                                                       data_gmic
gmake[2]: *** [CMakeFiles/libgmic.dir/build.make:76: CMakeFiles/libgmic.dir/src/gmic.cpp.o] Error 1
gmake[1]: *** [CMakeFiles/Makefile2:91: CMakeFiles/libgmic.dir/all] Error 2
gmake: *** [Makefile:136: all] Error 2
$ uname -rms
Linux 6.5.0-15-generic x86_64
```
