### 2024-06-09 GreycLab/gmic#53 @dtschump
Right,,the second one should read `libcgmic`.
