### 2024-06-09 GreycLab/gmic#53 @dtschump
Fixed with https://github.com/GreycLab/gmic/commit/bd4aac57bdad4a5e237d2655d81736575a4c9b2b
