### 2024-06-09 GreycLab/gmic#54 @dtschump
Good idea, I will add a CONTRIBUTING.md file ASAP. Thanks.
