### 2024-06-09 GreycLab/gmic#55 @dtschump
Thanks for reporting.
This works here for me, with latest development version 3.4.0:
~~~
~/work/src/gmic/build$ make
make: *** No targets specified and no makefile found.  Stop.
~/work/src/gmic/build$ cmake ..
-- The CXX compiler identification is GNU 11.4.0
-- The C compiler identification is GNU 11.4.0
-- Detecting CXX compiler ABI info
-- Check for working CXX compiler: /usr/bin/c++ - skipped
-- Detecting CXX compile features
-- Detecting C compiler ABI info
-- Detecting C compiler ABI info - done
-- Check for working C compiler: /usr/bin/cc - skipped
-- Detecting C compile features
-- Detecting C compile features - done
-- Found PkgConfig: /usr/bin/pkg-config (found version "0.29.2")
-- Found CImg: /home/tschumperle/work/src/gmic/src/CImg.h
-- Found OpenMP_C: -fopenmp (found version "4.5")
-- Found OpenMP_CXX: -fopenmp (found version "4.5")
-- Found OpenMP: TRUE (found version "4.5")
-- Found ZLIB: /usr/lib/x86_64-linux-gnu/libz.so (found version "1.2.11")
-- Found CURL: /usr/lib/x86_64-linux-gnu/libcurl.so (found version "7.81.0")
-- Found X11: /usr/include
-- Looking for XOpenDisplay in /usr/lib/x86_64-linux-gnu/libX11.so;/usr/lib/x86_64-linux-gnu/libXext.so
-- Looking for XOpenDisplay in /usr/lib/x86_64-linux-gnu/libX11.so;/usr/lib/x86_64-linux-gnu/libXext.so - found
-- Looking for gethostbyname
-- Looking for gethostbyname - found
-- Looking for connect
-- Looking for connect - found
-- Looking for remove
-- Looking for remove - found
-- Looking for shmat
-- Looking for shmat - found
-- Looking for IceConnectionNumber in ICE
-- Looking for IceConnectionNumber in ICE - found
-- Checking for module 'fftw3>=3.0'
--   Found fftw3, version 3.3.8
-- Found Fftw: fftw3
-- Checking for module 'GraphicsMagick++'
--   No package 'GraphicsMagick++' found
-- Could NOT find GraphicsMagick (missing: GRAPHICSMAGICK_LIBRARIES)
-- Found TIFF: /usr/lib/x86_64-linux-gnu/libtiff.so (found version "4.3.0")
-- Found PNG: /usr/lib/x86_64-linux-gnu/libpng.so (found version "1.6.37")
-- Found JPEG: /usr/lib/x86_64-linux-gnu/libjpeg.so (found version "80")
-- Checking for module 'OpenEXR'
--   Found OpenEXR, version 2.5.7
-- Found OpenEXR: IlmImf-2_5;Imath-2_5;Half-2_5;Iex-2_5;IexMath-2_5;IlmThread-2_5
-- Looking for pthread.h
-- Looking for pthread.h - found
-- Performing Test CMAKE_HAVE_LIBC_PTHREAD
-- Performing Test CMAKE_HAVE_LIBC_PTHREAD - Success
-- Found Threads: TRUE
-- Found GMicStdlib: /home/tschumperle/work/src/gmic/src/gmic_stdlib_community.h
-- The following OPTIONAL packages have been found:

 * OpenMP
 * CURL
 * X11
 * Fftw
 * TIFF
 * ZLIB
 * PNG
 * JPEG
 * PkgConfig
 * OpenEXR
 * Threads

-- The following REQUIRED packages have been found:

 * CImg
 * GMicStdlib

-- The following OPTIONAL packages have not been found:

 * GraphicsMagick

-- Configuring done
-- Generating done
-- Build files have been written to: /home/tschumperle/work/src/gmic/build
~/work/src/gmic/build$ make
[ 12%] Building CXX object CMakeFiles/libgmic.dir/src/gmic.cpp.o
In file included from /usr/include/stdio.h:894,
                 from /usr/include/c++/11/cstdio:42,
                 from /usr/include/c++/11/ext/string_conversions.h:43,
                 from /usr/include/c++/11/bits/basic_string.h:6608,
                 from /usr/include/c++/11/string:55,
                 from /usr/include/c++/11/bits/locale_classes.h:40,
                 from /usr/include/c++/11/locale:39,
                 from /home/tschumperle/work/src/gmic/src/gmic.h:185,
                 from /home/tschumperle/work/src/gmic/src/gmic.cpp:1786:
In function 'int fprintf(FILE*, const char*, ...)',
    inlined from 'static double gmic_library::gmic_image<T>::_cimg_math_parser::mp_vector_print(gmic_library::gmic_image<T>::_cimg_math_parser&) [with T = float]' at /home/tschumperle/work/src/gmic/src/CImg.h:29144:25:
/usr/include/x86_64-linux-gnu/bits/stdio2.h:105:24: warning: '%s' directive argument is null [-Wformat-overflow=]
  105 |   return __fprintf_chk (__stream, __USE_FORTIFY_LEVEL - 1, __fmt,
      |          ~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  106 |                         __va_arg_pack ());
      |                         ~~~~~~~~~~~~~~~~~
[ 25%] Linking CXX shared library libgmic.so
[ 25%] Built target libgmic
[ 37%] Building CXX object CMakeFiles/libgmicstatic.dir/src/gmic.cpp.o
In file included from /usr/include/stdio.h:894,
                 from /usr/include/c++/11/cstdio:42,
                 from /usr/include/c++/11/ext/string_conversions.h:43,
                 from /usr/include/c++/11/bits/basic_string.h:6608,
                 from /usr/include/c++/11/string:55,
                 from /usr/include/c++/11/bits/locale_classes.h:40,
                 from /usr/include/c++/11/locale:39,
                 from /home/tschumperle/work/src/gmic/src/gmic.h:185,
                 from /home/tschumperle/work/src/gmic/src/gmic.cpp:1786:
In function 'int fprintf(FILE*, const char*, ...)',
    inlined from 'static double gmic_library::gmic_image<T>::_cimg_math_parser::mp_vector_print(gmic_library::gmic_image<T>::_cimg_math_parser&) [with T = float]' at /home/tschumperle/work/src/gmic/src/CImg.h:29144:25:
/usr/include/x86_64-linux-gnu/bits/stdio2.h:105:24: warning: '%s' directive argument is null [-Wformat-overflow=]
  105 |   return __fprintf_chk (__stream, __USE_FORTIFY_LEVEL - 1, __fmt,
      |          ~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  106 |                         __va_arg_pack ());
      |                         ~~~~~~~~~~~~~~~~~
[ 50%] Linking CXX static library libgmic.a
[ 50%] Built target libgmicstatic
[ 62%] Building CXX object CMakeFiles/gmic.dir/src/gmic_cli.cpp.o
[ 75%] Linking CXX executable gmic
[ 75%] Built target gmic
[ 87%] Generating man/gmic.1
[gmic]./ Start G'MIC interpreter (v.3.4.0).
[gmic]./ Input custom command file '/home/tschumperle/work/src/gmic/src/gmic_stdlib.gmic' (2443 replaced, total: 4930).
[gmic]./ Input text-data file '/home/tschumperle/work/src/gmic/src/gmic_stdlib.gmic'.
[gmic]./ End G'MIC interpreter.
[ 87%] Built target man
[100%] Generating resources/gmic_bashcompletion.sh
[gmic]./ Start G'MIC interpreter (v.3.4.0).
[gmic]./ Input custom command file '/home/tschumperle/work/src/gmic/src/gmic_stdlib.gmic' (2443 replaced, total: 4930).
[gmic]./ Input text-data file '/home/tschumperle/work/src/gmic/src/gmic_stdlib.gmic'.
[gmic]./ Parse '#@cli' command(s) '*' and output in 'bashcompletion' mode.
[gmic]./ End G'MIC interpreter.
[100%] Built target bashcompletion
~~~

Maybe it could be due to the fact I've recently (yesterday) switched to a dev version 3.4.0 (instead of 3.3.7) ?
