### 2024-06-09 23ee5c8
Add 'CONTRIBUTING.md' file, for issue #54.
