### 2024-06-12 GreycLab/gmic#55 @Smattr
Any suggestions to debug/fix this? This is a pretty vanilla Ubuntu install I’m on.
