### 2024-06-16 GreycLab/gmic#55 @dtschump
Hello, sorry for the delay. I've been managing the release of the latest version (3.4.0) of G'MIC these last days.
Compiling G'MIC, assuming all required dependencies have been installed, should be as simple as that:
~~~
$  wget https://gmic.eu/files/source/gmic_3.4.0.tar.gz && tar zxvf gmic_3.4.0.tar.gz && cd gmic-3.4.0/src && make cli
~~~
(change `cli` by `gimp` to compile the G'MIC-Qt plug-in for GIMP).

Required dependencies are listed here : https://gmic.eu/download.html#linux

Hope this helps,

D.
