### 2024-06-16 GreycLab/gmic#55 @Smattr
So for the JOSS review we should base our assessment on the tarball rather than the Github repository?
