### 2024-06-17 GreycLab/gmic#55 @dtschump
> So for the JOSS review we should base our assessment on the tarball rather than the Github repository?

Yes, definitely.
the repository is just a work-in-progress version, which isn't the easiest way to get the latest version of the project and compile it. I'm going to update the download page to clarify this.
