### 2024-06-17 #6618 @Smattr
FYI for other reviewers, as per https://github.com/GreycLab/gmic/issues/55#issuecomment-2172312110 apparently we’re supposed to review the release tarball not the current content on Github.
