### 2024-06-17 #6618 @dtschump
Just a word on this:
The github repository is a "working repository" intended to contain the WIP source code. It contains two branches 'main' and 'develop', but the 'main' branches does not correspond to the state of the latest release.
Additionnally, the source code of G'MIC requires additionnal files that are **not** stored in the gmic repository, so using the code directly from the repository is definitely less practical if you only want to compile (or review) the project code.
