### 2024-06-17 #6618 @dtschump
I've updated the "Download" page from the website to reflect this : https://gmic.eu/download.html
