### 2024-06-25 GreycLab/gmic#54 @jamesrhester
That update looks sufficient to satisfy the JOSS guidelines to me.
