### 2024-06-25 #6618 @jamesrhester
Just a note on my approach to reviewing this:
1. verifying all of the functionality is impractical because there's so much, so I've simply made sure that a few of the demos and sample scripts from the repo work. May have played the 2048 game for longer than strictly necessary.
2. likewise it is impractical to check all of the supported platforms. As I run Debian, I've installed from a Debian mirror and used the fact that the Debian packager compiles from source to assess that it compiles and installs "by proxy".
