### 2024-07-14 GreycLab/gmic#60 @Smattr
[JOSS] mutex failure when building GMIC 3.4.0
Following https://github.com/GreycLab/gmic/issues/55, my next attempt at building from the release tarball:

```
$ wget https://github.com/GreycLab/gmic/releases/download/v.3.4.0/gmic_3.4.0.tar.gz
…
$ tar xvf gmic_3.4.0.tar.gz
…
$ cd gmic-3.4.0
$ cmake -DCMAKE_INSTALL_PREFIX=$(mktemp -d) -B build -S .
…
-- The following OPTIONAL packages have been found:
                                                                                                                                                                            * OpenMP
 * ZLIB
 * Threads

-- The following REQUIRED packages have been found:

 * CImg
 * GMicStdlib

-- The following OPTIONAL packages have not been found:

 * CURL
 * X11
 * Fftw
 * GraphicsMagick
 * TIFF
 * PNG
 * JPEG
 * PkgConfig
 * OpenEXR

-- Configuring done
-- Generating done
-- Build files have been written to: /tmp/tmp.GwIM5z52ev/gmic-3.4.0/build
$ cmake --build build
…
/usr/include/x86_64-linux-gnu/bits/stdio2.h:105:24: warning: \u2018%s\u2019 directive argument is null [-Wformat-overflow=]
  105 |   return __fprintf_chk (__stream, __USE_FORTIFY_LEVEL - 1, __fmt,
      |          ~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  106 |                         __va_arg_pack ());
      |                         ~~~~~~~~~~~~~~~~~
[ 50%] Linking CXX static library libgmic.a
[ 50%] Built target libgmicstatic
[ 62%] Building CXX object CMakeFiles/gmic.dir/src/gmic_cli.cpp.o
[ 75%] Linking CXX executable gmic
[ 75%] Built target gmic
[ 87%] Generating man/gmic.1
gmic: pthread_mutex_lock.c:94: ___pthread_mutex_lock: Assertion `mutex->__data.__owner == 0' failed.
Aborted (core dumped)
gmake[2]: *** [CMakeFiles/man.dir/build.make:73: man/gmic.1] Error 134
gmake[2]: *** Deleting file 'man/gmic.1'
gmake[1]: *** [CMakeFiles/Makefile2:169: CMakeFiles/man.dir/all] Error 2
gmake: *** [Makefile:136: all] Error 2
$ lsb_release -a
No LSB modules are available.
Distributor ID: Ubuntu
Description:    Ubuntu 22.04.4 LTS
Release:        22.04
Codename:       jammy
$ gcc --version
gcc (Ubuntu 11.4.0-1ubuntu1~22.04) 11.4.0
Copyright (C) 2021 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
```

`gmic` apparently does something incorrect with mutexes when generating the Bash completions. Any suggestions?
