### 2024-07-14 #6618 @Smattr
Taking another run at this just now, I still can’t get this project to build. https://github.com/GreycLab/gmic/issues/60. Unless I can compile this software, I’m not sure how I can review it.
