### 2024-07-21 GreycLab/gmic#60 @dtschump
Hello. Thanks for reporting this issue.
I've tested and it's working here for me (using Ubuntu 22.04 LTS and g++ 11.4.0).
Could you please give me some details on which distribution you encounter this error ?

As a work-around, I'd suggest to use the custom Makefile that is already present in folder `src/`, just as described in the G'MIC Download page : https://gmic.eu/download.html#linux

Thanks.
