### 2024-07-22 GreycLab/gmic#60 @Smattr
This is on a vanilla Ubuntu 22.04 Desktop with the same compiler. This is a VM with an almost out-of-the-box Ubuntu 22.04, upgraded to the latest packages via `apt`, so I’m not really sure what could be going wrong.

I’ll try the custom Makefile and report back…
