### 2024-08-11 GreycLab/gmic#60 @Smattr
Sorry for the delay on this. Following the instructions on that page, I still can’t get a successful build though the problem I hit is different:

```
$ sudo apt install git build-essential libgimp2.0-dev libcurl4-openssl-dev libfftw3-dev libtiff-dev libjpeg-dev libopenexr-dev qtbase5-dev qttools5-dev-tools
…
$ wget https://gmic.eu/files/source/gmic_3.4.0.tar.gz && tar zxvf gmic_3.4.0.tar.gz && cd gmic-3.4.0/src
…
$ make all
…
**
** Warning: folder 'zart' was not found !
** It is mandatory to compile 'zart'.
** Trying to retrieve it, with: $ cd ../../ && git clone https://github.com/c-koi/zart.git
**
Cloning into 'zart'...
remote: Enumerating objects: 735, done.
remote: Total 735 (delta 102), reused 99 (delta 49), pack-reused 572
Receiving objects: 100% (735/735), 818.15 KiB | 7.05 MiB/s, done.
Resolving deltas: 100% (550/550), done.
**
** Success !
**
cd ../zart && qmake CONFIG+=release GMIC_PATH=/tmp/tmp.AgN8qlxPkW/gmic-3.4.0/src zart.pro && make && strip zart
Info: creating stash file /tmp/tmp.AgN8qlxPkW/zart/.qmake.stash
Project MESSAGE: GMIC_PATH is set (/tmp/tmp.AgN8qlxPkW/gmic-3.4.0/src)
Project ERROR: Cannot find OpenCV
make[1]: *** [Makefile:637: zart] Error 3
make[1]: Leaving directory '/tmp/tmp.AgN8qlxPkW/gmic-3.4.0/src'
make: *** [Makefile:374: all] Error 2
…
```

Thanks for all your help with this, but I’m going to step down from the JOSS review. We don’t seem to be making progress and I am quite time-poor, so I’d prefer to unblock the review by letting someone else take over who might be able to get things working.
