### 2024-08-11 #6618 @Smattr
Sorry for the delay once again with this. I had another attempt using the [alternate instructions from @dtschump](https://github.com/GreycLab/gmic/issues/60#issuecomment-2241558045). Unfortunately I still can’t get this project to build.

I’d like to request to step down as a reviewer. I’ve had 3+ attempts at compiling this project and been unsuccessful. @dtschump has been very helpful and FWIW the problems may be entirely on my side. I’m a C++ dev myself and could probably try and debug this but, to be frank, this is not what I signed up for. I think given my limited availability and blockers here, it’s better to let someone else take over who may be able to progress this.

I’d like to highlight that none of this is a judgement on the quality of G’MIC. From where I reached in my review, I have the impression it is a mature and impressive project, for which a JOSS submission is somewhat of an afterthought – JOSS would be enriched by the inclusion of G’MIC, but G’MIC seems already very successful independent of this.
