### 2024-08-13 #6618 @gkthiruvathukal
@Smattr Thanks for your input. I actually see this as a major blocker if you have not been able to get this to compile. Other users are likely to encounter endless frustration, which will limit the impact of the contribution. Please don't step down. We value the time of our reviewers. At this point, your input/review is complete, since you gave us a clear statement about what improvements are needed.

Although I fully appreciate @dtschump's efforts to be helpful, this is not a key criterion for recommending acceptance. Everything starts with the ability to compile and run the code.

@jamesrhester can you confirm that as part of your review that you were able to successfully compile the code? This is crucial for me to make a recommendation how to proceed.
