### 2024-08-14 #6618 @Smattr
I think [this comment](https://github.com/openjournals/joss-reviews/issues/6618#issuecomment-2187778443) from jamesrhester explains:

> I … used the fact that the Debian packager compiles from source to assess that it compiles and installs "by proxy".
