### 2024-08-15 #6618 @jamesrhester
Yes, @Smattr is correct. If Debian provides the software then I am satisfied that it can be independently compiled from source. The only [patch from Debian](https://udd.debian.org/patches.cgi?src=gmic&version=2.9.4-4) is to allow clang to be used as well.

If this is not satisfactory for JOSS happy to have a go at compiling it myself.
