### 2024-08-15 #6618 @gkthiruvathukal
@jamesrhester I am mostly in agreement here. If this package can be installed from Debian, this might be satisfactory. However, I would like to know whether it is installable from the central repository or a PPA.

And if anyone has been able to build it, I could also be satisfied with a video showing all steps being executed successfully or even an Asciinema screencast.

Lastly, even something like a Homebrew formula would do.
