### 2024-09-04 #6618 @stefantalpalaru
> even something like a Homebrew formula would do

How about a Gentoo ebuild, instead?

https://github.com/stefantalpalaru/gentoo-overlay/tree/master/media-gfx/gmic

Build log: https://gist.github.com/stefantalpalaru/b342401125af0b3329cf86108932f402
