### 2024-10-12 #6618 @gkthiruvathukal
@danielskatz Just to clarify, as editor, am I supposed to recommend accept, or can you take it from here?
