### 2024-10-13 #6618 @danielskatz
@gkthiruvathukal - you should follow the steps in https://joss.readthedocs.io/en/latest/editing.html#post-review  (If you don't have time, I *can* do this, but would prefer that you do it)
