### 2024-10-15 #6618 @gkthiruvathukal
@dtschump Please let me know when the above author tasks at https://github.com/openjournals/joss-reviews/issues/6618#issuecomment-2414070636 are completed. These two are necessary before I can recommend acceptance. For convenience, please check them off here and follow up with the required archive DOI and version:
