### 2024-10-15 #6618 @dtschump
Thanks. Some news I have to share with you:
- Sadly, Sébastien Fourey, one of the co-author of this article, and co-developer of G'MIC has passed away last week. That explains why I didn't speak up before. Everyone in the lab was very shocked.
- In fact, I'd like to be able to add a small paragraph of tribute to Sébastien in this paper before it can be published. I don't know whether this means that the review process needs to be lengthened or not. Please keep me informed.
- Last Friday, I released version 3.4.3 of G'MIC. I'll mention this latest version in the article.
- So I'm going to do a final proofreading and make these few changes, as soon as possible.

Thank you for your understanding.
