### 2024-10-15 c996aa0
Latest edits. Add 'In Memoriam' section before 'References'.
