+++ paper.bib
@@ -14,12 +14,12 @@
   date = {2023-01-14},
 }
 
-@software{cimg,
-  author={Tschumperlé, David},
-  title={CImg: {C}++ {T}emplate {I}mage {P}rocessing {T}oolkit},
-  url = {http://cimg.eu},
-  version = {3.3.1},
-  date = {2023-05-30},
+@book{cimg,
+  title={Digital Image Processing with {C}++: Implementing Reference Algorithms with the {C}Img {L}ibrary (1st {E}d.)},
+  author={Tschumperle, David and Tilmant, Christophe and Barra, Vincent},
+  year={2023},
+  publisher={CRC Press},
+  doi = {10.1201/9781003323693}
 }
 
 @software{gmic,
