### 2024-10-15 #6618 @dtschump
Hello.
After paper review, I propose these last edits : https://github.com/GreycLab/gmic/commit/c996aa0597935399b5fd988380851c4239fb1b83
Let me know if that is OK for you.

I've also created an archive on Zenodo : https://zenodo.org/records/13936920
with the latest version released, numbered **3.4.3**.

Let me know if you need something else. Thanks.
