### 2024-10-16 #6618 @gkthiruvathukal
@dtschump First, I'm deeply saddened to read about the loss of your colleague. Hopefully, we can honor him by seeing this article published soon.

Sadly, I had a similar situation happen to me for a presentation/tutorial at an international conference just two weeks ago. It was very hard for all of us to give this presentation after losing him but we were able to do so and honor his memory.

Rest assured, this is not a blocker for us to proceed with your JOSS submission. I don't expect it to lengthen the review process as long as you made a good-faith effort to address reviewer comments. I believe you have done so but will await their confirmation.

Please take care of my checklist by checking all items that are complete. I will work with you to bring this to closure!

To reiterate: I offer my heartfelt condolences to you and his family/friends.
