### 2024-10-16 #6618 @jamesrhester
My condolences also. Happy for @gkthiruvathukal to directly approve any additional tribute text.
