### 2024-11-26 #6618 @gkthiruvathukal
I'm not entirely sure how to handle the MISSING DOIs section in the report and will defer to my EIC(s) in the know.
