### 2024-11-26 #6618 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
