### 2024-11-26 #6618 @dtschump
Well, OK, I've read the proof, and all seem OK to me.
If someone else can validate I'll be grateful :)
