### 2024-11-26 #6618 @danielskatz
👋 @gkthiruvathukal

1. It appears @drj11 didn't finish their review. Is this right? If so, please remove them from the reviewer list.
2. The checklist for @Smattr still has a number of unchecked items. I can't tell if they have finished their review or not.

Thanks
