### 2024-11-27 #6618 @Smattr
Just the blocker [discussed above](https://github.com/openjournals/joss-reviews/issues/6618#issuecomment-2282896548):

> Unfortunately I still can’t get this project to build.

My understanding was that you were considering jamesrhester’s experience sufficient and [were going to proceed with acceptance](https://github.com/openjournals/joss-reviews/issues/6618#issuecomment-2402233181):

> That would be acceptable. Based on the earlier feedback, I am satisfied that this package can be built successfuly and meets our criteria.
>
> I am ready to proceed toward acceptance and will follow up with a checklist shortly.
