### 2024-11-27 #6618 @danielskatz
@Smattr - please check off the items on [your checklist](https://github.com/openjournals/joss-reviews/issues/6618#issuecomment-2156229068) that you can. A number of these aren't dependent on installation of the software.

And as the track editor, I am not willing to accept this without two complete reviews.  If you can check off all the other items, we can see if we can find someone else who can test the build/install part, in addition to what @jamesrhester has already done.
