### 2024-12-03 #6618 @danielskatz
@Smattr - can you finish reviewing the paper and checking off the items related to it?
