### 2024-12-04 #6618 @Smattr
Sorry for the delay. I’ve progressed my review just now.

I haven’t ticked the tests box though I notice @jamesrhester has. James, can you point me to the tests? I couldn’t find any automated or manual tests, nor wording about them.
