### 2024-12-05 #6618 @jamesrhester
> Sorry for the delay. I’ve progressed my review just now.
>
> I haven’t ticked the tests box though I notice @jamesrhester has. James, can you point me to the tests? I couldn’t find any automated or manual tests, nor wording about them.

I ran some of the example commands at https://gmic.eu/reference/examples_of_use.html#top as well as other commands in that documentation to reassure myself that the software performed as expected. My understanding of this checklist point is to make sure that the software works as advertised. I don't think we are meant to require that there are formal tests that can be run that automatically verify correct functioning.
