### 2024-12-05 #6618 @Smattr
I guess I’m not sure what to do about my remaining item then. I read it:

> **Automated tests**: Are there automated tests or manual steps described so that the functionality of the software can be verified?

as requiring a test suite or instructions on tests to manually run. Though given I can’t build the software, I won’t be able to verify the tests pass, so perhaps there is limited value in this item anyway.
