### 2024-12-05 #6618 @dtschump
Just a quick note on automated tests: if you have the software installed (command line tool), then
~~~
$ gmic input_text https://gmic.eu/gmic_stdlib.\$_version parse_cli images
~~~
runs a kind of automated test.

Here on my machine, it displays this:
~~~
[gmic]./ Start G'MIC interpreter (v.3.5.0).
[gmic]./ Input text-data file 'https://gmic.eu/gmic_stdlib.350'.
[gmic]./ Parse '#@cli' command(s) '*' and output in 'images' mode.

 ** Section #1: Global Options.

 ** Section #2: Input / Output.

[1] Command 'command': $ image.jpg command "foo : mirror y deform $""1" +foo[0] 5 +foo[0] 15 (done in 0.04s).
[2] Command 'display_clut': $ clut tealorange clut summer clut 60s display_clut 400 (done in 1.55s).
[3] Command 'display_fft': $ image.jpg +display_fft (done in 0.13s).
[4] Command 'display_graph': $ 128,1,1,1,'cos(x/10+u)' +display_graph 400,300,3 (done in 0.02s).
[5] Command 'display_histogram': $ image.jpg +display_histogram 512,300 (done in 0.03s).
[6] Command 'display_parametric': $ 1024,1,1,2,'t=x/40;(!c?sin(t):cos(t))*(exp(cos(t))-2*cos(4*t)-sin(t/12)^5)' display_parametric 512,512 (done in 0.02s).
[7] Command 'display_parametric': $ 1000,1,1,2,u(-100,100) quantize 4,1 noise 12 channels 0,2 +normalize 0,255 append c display_parametric 512,512,0.1,8 (done in 0.05s).
[8] Command 'display_polar': $ 300,1,1,1,'0.3+abs(cos(10*pi*x/w))+u(0.4)' display_polar 512,512,4,200,255,200 (done in 0.02s).
[9] Command 'display_polar': $ 3000,1,1,1,'x^3/1e10' display_polar 400,400,1,-1,,,0,{15*360} (done in 0.02s).
[10] Command 'display_quiver': $ image.jpg +luminance gradient[-1] xy rv[-2,-1] *[-2] -1 a[-2,-1] c crop 60,10,90,30 +display_quiver[1] , (done in 0.04s).
[11] Command 'display_rgba': $ image.jpg +norm threshold[-1] 40% blur[-1] 3 normalize[-1] 0,255 append c display_rgba (done in 0.02s).
[12] Command 'display_tensors': $ image.jpg +diffusiontensors 0.1,0.9 rescale2d. 64 +display_tensors. 16,2 (done in 0.1s).
[13] Command 'display_warp': $ 400,400,1,2,'x=x-w/2;y=y-h/2;r=sqrt(x*x+y*y);a=atan2(y,x);5*sin(r/10)*[cos(a),sin(a)]' +display_warp 10 (done in 0.01s).
[14] Command 'font': $ 400,300,1,3 text "Hello World!",0.5~,0.5~,${"font \"Cheese Burger\",80"},1,255,255,128 (done in 0.3s).
[15] Command 'function1d': $ function1d 1,0,0,10,30,40,20,70,30,80,0 +display_graph 400,300 (done in 0.01s).
[16] Command 'identity': $ identity 5,1 identity 8,8 (done in 0s).
[17] Command 'input': $ input image.jpg (done in 0.01s).
[18] Command 'input': $ input (1,2,3;4,5,6;7,8,9^9,8,7;6,5,4;3,2,1) (done in 0s).
[19] Command 'input': $ image.jpg (1,2,3;4,5,6;7,8,9) (255^128^64) 400,400,1,3,'(x>w/2?x:y)*c' (done in 0.01s).
[20] Command 'pass': $ command "average : pass$""1 add[^-1] [-1] remove[-1] div 2" sample ? +mirror y +average[0] [1] (done in 0.03s).
[21] Command 'random_pattern': $ repeat 6 { random_pattern 256 } (done in 0.92s).
[22] Command 'serialize': $ image.jpg +serialize uint8 +unserialize[-1] (done in 0.08s).
[23] Command 'shape_circle': $ shape_circle , (done in 0.01s).
[24] Command 'shape_cupid': $ shape_cupid , (done in 0.03s).
[25] Command 'shape_diamond': $ shape_diamond , (done in 0s).
[26] Command 'shape_dragon': $ shape_dragon , (done in 0.15s).
[27] Command 'shape_fern': $ shape_fern , (done in 0.05s).
[28] Command 'shape_gear': $ shape_gear , (done in 0s).
[29] Command 'shape_heart': $ shape_heart , (done in 0.01s).
[30] Command 'shape_menger': $ shape_menger 4 surfels3d , color3d 200 m3d 3 (done in 0.35s).
[31] Command 'shape_mosely': $ shape_mosely 4 surfels3d , color3d 200 m3d 3 (done in 0.23s).
[32] Command 'shape_polygon': $ repeat 6 { shape_polygon 256,{3+$>} } (done in 0s).
[33] Command 'shape_rays': $ shape_rays 400,50%,50%,7 shape_rays 400,50%,50%,3,0,3 (done in 0.01s).
[34] Command 'shape_snowflake': $ repeat 6 { shape_snowflake 256,$> } (done in 0.03s).
[35] Command 'shape_star': $ repeat 9 { shape_star 256,{$>+2} } (done in 0.02s).
[36] Command 'shared': $ image.jpg shared 1 blur[-1] 3 remove[-1] (done in 0.01s).
[37] Command 'shared': $ image.jpg repeat s { shared 25%,75%,0,$> mirror[-1] x remove[-1] } (done in 0.01s).
[38] Command 'sample': $ repeat 6 { sample } (done in 0.12s).
[39] Command 'store': $ sample eagle,earth store img1,img2 input $img2 $img1 (done in 0.03s).
[40] Command 'testimage2d': $ testimage2d 512 (done in 0.03s).
[41] Command 'uniform_distribution': $ uniform_distribution 64,3 * 255 +distribution3d circles3d[-1] 10 (done in 0s).
(...)
~~~

It does not verify that the result of each command is "correct" (by comparing the result of the applied command to a reference image), but at least it verifies that all G'MIC commands run without any issue.
