### 2024-12-05 #6618 @Smattr
Not sure what the bar is here, but from my perspective this checkbox would be ticked by adding the above to the repo readme.
