### 2024-12-05 4f01196
Add a description of the commands to verify the correctness of the software, in README and README.md.
