### 2024-12-05 #6618 @dtschump
@Smattr Good idea. I've added a short section in files README and README.md

https://github.com/GreycLab/gmic/commit/4f0119624563d1daa45a038cfaf10678af47ee26
