### 2024-12-05 #6618 @danielskatz
👋 @dtschump and @Smattr - Is there any chance you would be able to talk offline to work out the issue that @Smattr is having in installing the software? This would help the JOSS review, and perhaps would also help the software.

@Smattr - if you don't have time to do this, that's ok, but we will have then find another reviewer, so I wanted to see if this was possible on your side before going through this process.
