### 2024-12-05 #6618 @dtschump
@danielskatz Of course, but I'm not an expert in compiling software (in general) on other distros than Debian/Ubuntu/Windows though. I know this is possible to compile G'MIC on a Mac (some people succeeded in doing so), but I'm afraid I can't be of much help if that the targeted platform for @Smattr .
