### 2024-12-06 #6618 @Smattr
I’m going to have to respectfully decline. @dtschump has been extremely helpful in getting me past various blockers. But, to quote my prior message:

> I’m a C++ dev myself and could probably try and debug this but, to be frank, this is not what I signed up for. I think given my limited availability and blockers here, it’s better to let someone else take over who may be able to progress this.

If it’s preferable to replace me with another reviewer, please go ahead.
