### 2024-12-06 #6618 @danielskatz
I would add one of them (who is willing) as a reviewer, have them generate a checklist, and if they want to complete everything, that would be fine, but really all we need is them to check the functionality part. The published paper would then list the three reviewers.
