### 2024-12-17 #6618 @NicholasSynovic
I successfully built the software for the latest release of G'MIC (3.4.3) on Pop OS 22.04. However, I was unable to test the functionality of it with the provided commands:

1. `gmic input_text https://gmic.eu/gmic_stdlib.\$_version parse_cli images`
2. `gmic input_text https://gmic.eu/gmic_stdlib.\$_version parse_gui images`

These commands were found at the bottom of the README.md file on the GitHub page as of 12/16/2024

For command 1, I got the following error:

```
[3] Command 'display_fft': $ image.jpg +display_fft
[gmic] *** Error in ./parse_cli/parse_cli_images/*repeat/(...)/_display_fft/*foreach/fftpolar/*foreach/ *** Command 'fft': gmicList<float32>::FFT(): Specified real and imaginary parts (640,427,1,3) have non 2^N dimension along the Y-axis.
[gmic] Command 'fft' has the following description:

  fft (+):
      _{ x | y | z }...{ x | y | z }

    Compute the direct fourier transform (real and imaginary parts) of selected images,
    optionally along the specified axes only.
    See also: ifft.

    Example:
      [#1] image.jpg luminance +fft append[-2,-1] c norm[-1] log[-1] shift[-1] 50%,50%,0,0,2
      [#2] image.jpg w2:=int(w/2) h2:=int(h/2) fft shift $w2,$h2,0,0,2 ellipse $w2,$h2,30,30,0,1,0 shift -$w2,-$h2,0,0,2 ifft remove[-1]

    Tutorial: https://gmic.eu/oldtutorial/_fft
```

For command 2, I got the following error:

```
[gmic]./ Start G'MIC interpreter (v.3.4.3).
[gmic]./ Input text-data file 'https://gmic.eu/gmic_stdlib.343'.
[gmic]./ Parse '#@gui' filters '*' and output in 'images' mode.
  > Parsing done!
  >> Generate output, in 'images' mode.
  [#1] About G'MIC:
  [#2] Download External Data: gui_no_preview 0
  [#3] Gmicky - Roddy: fx_gmicky_preview 0
  [#4] Privacy Notice:
  [#5] Release Notes:
  [#6] Tribute to Sébastien:
  [#7] What's New?:
  [#8] Array [Faded]: fx_array_fade_preview 2,2,0,0,80,90,0,0
  [#9] Array [Mirrored]: fx_array_mirror_preview 1,0,0,2,0,0,0
  [#10] Array [Random]: array_random 5,5,7,7
  [#11] Array [Random Colors]: fx_array_color 5,5,0.5
  [#12] Array [Regular]: fx_array_preview 2,2,0,0,0,0
  [#13] Ascii Art: fx_asciiart_preview 5," .oO0",16,15,16,2,0,0.2,0,0,"","gmic_asciiart.txt"
  [#14] Chessboard: fx_chessboard_preview 64,64,0,0,0,0.25,0,0,0,255,255,255,255,255,0,50,50
  [#15] Dices: fx_dices 2,24,1
  [#16] Drawn Montage: fx_drawn_montage_preview 0,0,0,0,-10,50,50,0,"-1","0:0:0:0:50:50:0","0:0:0:0:50:50:0","0:0:0:0:50:50:0","0:0:0:0:50:50:0","0:0:0:0:50:50:0","0:0:0:0:50:50:0","0:0:0:0:50:50:0","0:0:0:0:50:50:0","0:0:0:0:50:50:0","0:0:0:0:50:50:0","0:0:0:0:50:50:0","0:0:0:0:50:50:0","0:0:0:0:50:50:0","0:0:0:0:50:50:0","0:0:0:0:50:50:0","0:0:0:0:50:50:0"
  > 2/2 Extract Objects: fx_extract_objects_preview 0,0,20,50,0.3,0,0,1
  [#18] Grid [Cartesian]: fx_imagegrid 10,10
  [#19] Grid [Hexagonal]: fx_imagegrid_hexagonal 32,0.1,1
  [#20] Grid [Triangular]: fx_imagegrid_triangular 10,18,0,0,0,0,128
  [#21] Loose Photos: fx_loose_photos_preview 60,40,50,100,50,360,0,2,255,255,255,50,25,0,0,0,0,0,0,50,1,1,1
  [#22] Make Seamless [Diffusion]: fx_make_seamless_preview 0,0,3,0,50,50
[gmic] *** Error in ./parse_gui/parse_gui_images/*repeat/(...)/*if/gui_error_preview/gui_print_preview/drgba/ *** Command 'to_rgb': Image [0] is not a G,GA,RGB or RGBA image (0 channels).
```

Does this error occur with any other reviewers? If so, how do I remedy the situation?
