### 2024-12-17 #6618 @dtschump
Hello @NicholasSynovic
Concerning your errors :
- It seems the build you have done has not enabled the use of the **FFTW** library natively, which means that `gmic` uses then a "default" FFT (Fast Fourier Transform) routine, which works only when your image has a dimension that is a power of 2 (which is not the case with the example image used for the tests).
This is a big limitation, because every filter that requires Fast Fourier Transform won't work correctly with images that does not respect this constraint (this is the case here, with the filter `Make Seamless [Diffusion]`.

So, could you please try installing the FFTW library and corresponding development files, so that  `cmake` can find it when compiling `gmic`?
It should display something like when invoking `cmake`:
~~~
-- The following OPTIONAL packages have been found:

 * OpenMP
 * CURL
 * X11
 * Fftw
 * TIFF
 * ZLIB
 * PNG
 * JPEG
 * PkgConfig
 * OpenEXR
 * Threads
~~~

Thanks!
