### 2024-12-20 #6618 @NicholasSynovic
Hi @dtschump , I was able to compile G'MIC with FFTW in an Ubuntu Jammy (22.04) Docker container. I did this because FFTW was not being detected by G'MIC on my host OS (I was unable to diagnose why exactly; most likely due to shell configuration).

I followed the build instructions for Linux listed here: [https://gmic.eu/download.html](https://gmic.eu/download.html). However, I had to also install the `libwebp-dev` package from APT within my container. This dependency is not listed within the instructions. Without this dependency, the software was unable to build with `make cli`.

I was able to successfully test `gmic input_text https://gmic.eu/gmic_stdlib.\$_version parse_cli images`.

I was able to successfully test `gmic input_text https://gmic.eu/gmic_stdlib.\$_version parse_cli images`.

As I am now able to successfully build and test the software, I belive that we can move on with the rest of the review.
