### 2024-12-21 #6618 @gkthiruvathukal
@danielskatz I think @NicholasSynovic mentioned that the documentation needs to be updated to address the additional `libwebp-dev` package.

@dtschump Can you please check this and let us know when it is done?

Thanks, @NicholasSynovic for your help on such short notice, especially with the end of semester!
