### 2024-12-21 2f69566
Add 'libwebp-dev' dependency in instructions to compile from source on Linux.
