### 2024-12-21 #6618 @dtschump
I've added the `libwebp-dev` dependency in the documentation (instructions to compile from source on Linux) , at https://gmic.eu/download.html#linux.

Corresponding commit:
https://github.com/GreycLab/gmic/commit/2f69566b03fd13af4414d944f44857bf876b3ce4
