### 2025-01-09 #6618 @gkthiruvathukal
@dtschump Looks like we have a lingering issue here for the DOI 10.1167/16.12.326. Thoughts?
