### 2025-01-09 8228777
JOSS: Article 'A neural algorithm of artistic style' now cited as the journal version in the .bib file, rather than the ArXiv preprint.
