+++ paper.bib
@@ -190,8 +190,12 @@
 @article{gatys2015,
   title={A neural algorithm of artistic style},
   author={Gatys, Leon A and Ecker, Alexander S and Bethge, Matthias},
-  journal={arXiv preprint arXiv:1508.06576},
-  year={2015}
+  journal={Journal of Vision},
+  volume={16},
+  number={12},
+  pages={326},
+  doi={10.1167/16.12.326},
+  year={2016}
 }
 
 @article{ray2023outflows,
