### 2025-01-09 #6618 @dtschump
Hello.
https://github.com/GreycLab/gmic/commit/82287776c3cb7502476ff743eac1d49fe113e747 fixes this issue. The article listed in the .bib file was indeed the ArXiv preprint, not the journal version (for which the DOI is proposed).
In fact, this makes no differences because this citation has been remove from the paper (when I was asked to reduce the number of pages).
I suppose the editorialbot script does not check that all references listed in the .bib files are indeed cited in the paper.
I think it's interesting to keep those 'uncited' publications in the .bib files, as they are still related to the content of the paper, even if they do not appear explicitely in it.
