### 2025-01-09 #6618 @danielskatz
Yes, editorialbot checks the .bib file and doesn't cross-check which references are cited.
