### 2025-01-09 #6618 @danielskatz
I'll check the proof of this in the next day or two, and then we can proceed.
