### 2025-01-09 #6618 @danielskatz
@dtschump - I've suggested a number of changes in https://github.com/GreycLab/gmic/pull/94, and also note that you don't have an acknowledgments section, which could be used to mention funding, if needed.

Can you merge the changes, or let me know what you disagree with?
And then consider adding an acknowledgments section?
