### 2025-01-09 d522b0f
Add an 'Acknowledgments' section to the JOSS paper.
