### 2025-01-09 #6618 @dtschump
Hello @danielskatz .
I've merged your changes, thanks a lot for the corrections.

I've also added an "Acknowledgements" section, as suggested, with commit: https://github.com/GreycLab/gmic/commit/d522b0fb6399c02212669afb7e67a4f05c6ed65a

Thanks for the suggestion!
