### 2025-01-09 #6618 @danielskatz
@dtschump - Can you check the "final" proof above and verify that it all looks good to you?
