### 2025-01-09 23e1178
Fix link to new G'MIC-Qt github repository.
