### 2025-01-09 #6618 @dtschump
@danielskatz , I've read the proof and found a single issue, that is fixed with: https://github.com/GreycLab/gmic/commit/23e1178d913c4ddc9f8ced09ba4025ecd248e272
(link to G'MIC-Qt github repo was not correct anymore, as it moved from Sebastien Fourey's account to the GREYC account, after his passing).
