### 2025-01-09 #6618 @danielskatz
Congratulations to @dtschump (David Tschumperlé) and co-authors on your publication!!

JOSS fully depends on volunteers and couldn't do this without you
