### 2025-01-09 #6618 @dtschump
Thanks @danielskatz and thanks to the reviewers! :fireworks:
