### 2025-01-10 #6618 @dtschump
@Smattr , In France, there's a saying that free software supporters often use: “The road is long, but the path is free”.
We all know that developing free software is all about patience. :)
